Pedidos libres

+++ public/script/editarProduccion.js
@@ -1,18 +1,18 @@
 let totalPedido = 0;
 
 function pedidoProduccionCalcImporte(e){
-    if(isNaN(parseInt(e.target.value))){
+    if(isNaN(parseFloat(e.target.value))){
         e.target.value = 0;
         return;
     }
     let costo = document.querySelector("#costo" + e.target.dataset.id);
-    costo = parseInt(costo.innerHTML);
+    costo = parseFloat(costo.innerHTML);
     if(costo === 0){
         mostrarError("El producto seleccionado no tiene en monto cargado, por favor comunicarse con administración para modificarlo");
         e.target.value = 0;
         return;
     }
-    document.querySelector(`#importe${e.target.dataset.id}`).innerHTML = costo * parseInt(e.target.value);
+    document.querySelector(`#importe${e.target.dataset.id}`).innerHTML = costo * parseFloat(e.target.value);
     pedidoProduccionCalcTotal();
 }
 
@@ -26,7 +26,7 @@ function pedidoProduccionCalcImportes(){
 
 function pedidoProduccionCalcTotal(){
     let montosArray = [];
-    document.querySelectorAll(".importe").forEach((monto) => montosArray.push(parseInt(monto.innerHTML)));
+    document.querySelectorAll(".importe").forEach((monto) => montosArray.push(parseFloat(monto.innerHTML)));
     totalPedido = montosArray.reduce((acumulador, valor) => {
         return acumulador + valor;
     }, 0);
